Encapsulate Swap into module
