+++ pallets/swap/Cargo.toml
@@ -6,7 +6,7 @@ edition = { workspace = true }
 [dependencies]
 alloy-primitives = { workspace = true }
 approx = { workspace = true }
-codec = {workspace = true }
+codec = { workspace = true }
 frame-support = { workspace = true }
 frame-system = { workspace = true }
 safe-math = { workspace = true }
@@ -16,6 +16,7 @@ sp-arithmetic = { workspace = true }
 sp-runtime = { workspace = true }
 sp-std = { workspace = true }
 substrate-fixed = { workspace = true }
+uuid = { workspace = true, features = ["v4"] }
 
 pallet-subtensor-swap-interface = { workspace = true }
 
@@ -37,4 +38,5 @@ std = [
 	"sp-runtime/std",
 	"sp-std/std",
 	"substrate-fixed/std",
+	"uuid/std",
 ]
