+++ pallets/swap/src/pallet/impls.rs
@@ -6,11 +6,14 @@ use sp_runtime::traits::AccountIdConversion;
 use substrate_fixed::types::U64F64;
 
 use super::pallet::*;
-use crate::tick::{Tick, TickIndex};
+use crate::{
+    NetUid, Position, PositionId, RemoveLiquidityResult,
+    tick::{Tick, TickIndex},
+};
 
 impl<T: Config> Pallet<T> {
-    // initializes V3 swap for a subnet
-    fn maybe_initialize_v3(netuid: u16) -> Result<(), Error<T>> {
+    // initializes V3 swap for a subnet if needed
+    fn maybe_initialize_v3(netuid: NetUid) -> Result<(), Error<T>> {
         if SwapV3Initialized::<T>::get(netuid) {
             return Ok(());
         }
@@ -34,16 +37,17 @@ impl<T: Config> Pallet<T> {
         // Set initial (protocol owned) liquidity and positions
         // Protocol liquidity makes one position from TickIndex::MIN to TickIndex::MAX
         // We are using the sp_arithmetic sqrt here, which works for u128
-        let liquidity = helpers_128bit::sqrt(tao_reserve as u128 * alpha_reserve as u128) as u64;
+        let liquidity =
+            helpers_128bit::sqrt((tao_reserve as u128).saturating_mul(alpha_reserve as u128))
+                as u64;
         let protocol_account_id = T::ProtocolId::get().into_account_truncating();
 
-        Self::add_liquidity(
+        let _ = Self::add_liquidity(
             netuid,
             &protocol_account_id,
             TickIndex::MIN,
             TickIndex::MAX,
             liquidity,
-            true,
         )?;
 
         Ok(())
@@ -67,14 +71,9 @@ impl<T: Config> Pallet<T> {
     /// - `tick_low`: The lower bound of the price tick range.
     /// - `tick_high`: The upper bound of the price tick range.
     /// - `liquidity`: The amount of liquidity to be added.
-    /// - `protocol`: A boolean flag indicating whether the operation is protocol-managed:
-    ///   - `true` -> Do not use this value outside of this implementation. Liquidity is added **without**
-    ///               withdrawing balances.
-    ///   - `false` -> Use this value for all user transactions. Liquidity is added
-    ///               **after withdrawing balances**.
     ///
     /// ### Returns:
-    /// - `Ok(u64)`: The final liquidity amount added.
+    /// - `Ok((u64, u64))`: (tao, alpha) amounts at new position
     /// - `Err(SwapError)`: If the operation fails due to insufficient balance, invalid tick range,
     ///   or other swap-related errors.
     ///
@@ -82,16 +81,15 @@ impl<T: Config> Pallet<T> {
     /// - [`SwapError::InsufficientBalance`] if the account does not have enough balance.
     /// - [`SwapError::InvalidTickRange`] if `tick_low` is greater than or equal to `tick_high`.
     /// - Other [`SwapError`] variants as applicable.
-    fn add_liquidity(
-        netuid: u16,
-        account_id: &<T as frame_system::Config>::AccountId,
+    pub fn add_liquidity(
+        netuid: NetUid,
+        account_id: &T::AccountId,
         tick_low: TickIndex,
         tick_high: TickIndex,
         liquidity: u64,
-        protocol: bool,
-    ) -> Result<(), Error<T>> {
+    ) -> Result<(u64, u64), Error<T>> {
         ensure!(
-            Positions::<T>::get(netuid, account_id).len() <= T::MaxPositions::get() as usize,
+            Self::count_positions(netuid, account_id) <= T::MaxPositions::get() as usize,
             Error::<T>::MaxPositionsExceeded
         );
 
@@ -105,16 +103,23 @@ impl<T: Config> Pallet<T> {
 
         Self::update_liquidity_if_needed(netuid, tick_low, tick_high, liquidity as i128);
 
-        // // New position
-        // let position = Position {
-        //     tick_low,
-        //     tick_high,
-        //     liquidity,
-        //     fees_tao: 0_u64,
-        //     fees_alpha: 0_u64,
-        // };
+        // New position
+        let position = Position {
+            id: PositionId::new(),
+            tick_low,
+            tick_high,
+            liquidity,
+            fees_tao: 0,
+            fees_alpha: 0,
+        };
 
-        // // If this is a user transaction, withdraw balances and update reserves
+        let current_price = AlphaSqrtPrice::<T>::get(netuid);
+        let (tao, alpha) = position.to_token_amounts(current_price)?;
+
+        // If this is a user transaction, withdraw balances and update reserves
+        // TODO this should be returned (tao, alpha) from this function to prevent
+        // mutation of outside storage - the logic should be passed to the user of
+        // pallet_subtensor_swap_interface
         // if !protocol {
         //     let current_price = self.state_ops.get_alpha_sqrt_price();
         //     let (tao, alpha) = position.to_token_amounts(current_price)?;
@@ -127,12 +132,11 @@ impl<T: Config> Pallet<T> {
         //     self.state_ops.set_alpha_reserve(new_alpha_reserve);
         // }
 
-        // // Create a new user position
-        // self.state_ops.create_position(account_id, position);
+        Positions::<T>::insert(&(netuid, account_id, position.id), position);
 
-        // SwapV3Initialized::<T>::set(netuid, true);
+        SwapV3Initialized::<T>::set(netuid, true);
 
-        Ok(())
+        Ok((tao, alpha))
     }
 
     /// Adds or updates liquidity at a specific tick index for a subnet
@@ -141,7 +145,7 @@ impl<T: Config> Pallet<T> {
     /// * `netuid` - The subnet ID
     /// * `tick_index` - The tick index to add liquidity to
     /// * `liquidity` - The amount of liquidity to add
-    fn add_liquidity_at_index(netuid: u16, tick_index: TickIndex, liquidity: u64, upper: bool) {
+    fn add_liquidity_at_index(netuid: NetUid, tick_index: TickIndex, liquidity: u64, upper: bool) {
         // Convert liquidity to signed value, negating it for upper bounds
         let net_liquidity_change = if upper {
             -(liquidity as i128)
@@ -165,14 +169,71 @@ impl<T: Config> Pallet<T> {
         });
     }
 
+    /// Remove liquidity and credit balances back to account_id
+    ///
+    /// Account ID and Position ID identify position in the storage map
+    pub fn remove_liquidity(
+        netuid: NetUid,
+        account_id: &T::AccountId,
+        position_id: PositionId,
+    ) -> Result<RemoveLiquidityResult, Error<T>> {
+        let Some(mut pos) = Positions::<T>::get((netuid, account_id, position_id)) else {
+            return Err(Error::<T>::LiquidityNotFound);
+        };
+        let current_tick_index = CurrentTickIndex::<T>::get(netuid);
+
+        // Collect fees and get tao and alpha amounts
+		// let (fee_tao, fee_alpha) = self.collect_fees(&mut pos);
+        //     let current_price: SqrtPrice = self.state_ops.get_alpha_sqrt_price();
+        //     let (tao, alpha) = pos.to_token_amounts(current_price)?;
+
+        //     // Update liquidity at position ticks
+        //     self.remove_liquidity_at_index(pos.tick_low, pos.liquidity, false);
+        //     self.remove_liquidity_at_index(pos.tick_high, pos.liquidity, true);
+
+        //     // Update current tick liquidity
+        //     if (pos.tick_low <= current_tick_index) && (current_tick_index <= pos.tick_high) {
+        //         let new_current_liquidity = self
+        //             .state_ops
+        //             .get_current_liquidity()
+        //             .saturating_sub(pos.liquidity);
+        //         self.state_ops.set_current_liquidity(new_current_liquidity);
+        //     }
+
+        //     // Remove user position
+        //     self.state_ops.remove_position(account_id, position_id);
+
+        //     // Deposit balances
+        //     self.state_ops.deposit_balances(account_id, tao, alpha);
+
+        //     // Update reserves
+        //     let new_tao_reserve = self.state_ops.get_tao_reserve().saturating_sub(tao);
+        //     self.state_ops.set_tao_reserve(new_tao_reserve);
+        //     let new_alpha_reserve = self.state_ops.get_alpha_reserve().saturating_sub(alpha);
+        //     self.state_ops.set_alpha_reserve(new_alpha_reserve);
+
+        //     // TODO: Clear with R&D
+        //     // Update current price (why?)
+        //     // self.state_ops.set_alpha_sqrt_price(sqrt_price);
+
+        //     // Return Ok result
+        //     Ok(RemoveLiquidityResult {
+        //         tao,
+        //         alpha,
+        //         fee_tao,
+        //         fee_alpha,
+        //     })
+        todo!()
+    }
+
     /// Gets the current tick index for a subnet, ensuring it's within valid bounds
-    fn bounded_current_tick_index(netuid: u16) -> TickIndex {
+    fn bounded_current_tick_index(netuid: NetUid) -> TickIndex {
         let current_price = AlphaSqrtPrice::<T>::get(netuid);
         TickIndex::from_sqrt_price_bounded(current_price)
     }
 
     /// Clamps the subnet's sqrt price when tick index is outside of valid bounds
-    fn clamp_sqrt_price(netuid: u16, tick_index: TickIndex) {
+    fn clamp_sqrt_price(netuid: NetUid, tick_index: TickIndex) {
         if tick_index >= TickIndex::MAX || tick_index <= TickIndex::MIN {
             let corrected_price = tick_index.to_sqrt_price_bounded();
             AlphaSqrtPrice::<T>::set(netuid, corrected_price);
@@ -186,7 +247,7 @@ impl<T: Config> Pallet<T> {
     /// liquidity parameter. It uses i128 to safely handle values up to u64::MAX in both positive
     /// and negative directions.
     fn update_liquidity_if_needed(
-        netuid: u16,
+        netuid: NetUid,
         tick_low: TickIndex,
         tick_high: TickIndex,
         liquidity: i128,
@@ -203,5 +264,17 @@ impl<T: Config> Pallet<T> {
                 }
             });
         }
+    }
+
+    /// Returns the number of positions for an account in a specific subnet
+    ///
+    /// # Arguments
+    /// * `netuid` - The subnet ID
+    /// * `account_id` - The account ID
+    ///
+    /// # Returns
+    /// The number of positions that the account has in the specified subnet
+    fn count_positions(netuid: NetUid, account_id: &T::AccountId) -> usize {
+        Positions::<T>::iter_prefix_values((netuid, account_id.clone())).count()
     }
 }
